Changes for more intensive testing of debian_recheck.py
debian_recheck.py: facilitate more detailed testing -
move literal paths/filenames to instance variables (avoid defining constants
in methods, and allow a test to change the variables to reference files
other than the ones present in the current system environment)

Because Travis runs under ubuntu, the current tests, when run under Travis,
just verify that DebianRecheck fails safe if it finds something that it
doesn't understand.

The new tests in test_debian_recheck.py substitute the default apt/dpkg
lookup files with ones collected from instances of the following:

- debian stretch (v9)
- debian buster (v10)
- devuan ascii (v2.1)
- devuan beowulf (v3.0)

(Note that because many of the files are under /var/lib in the originals,
problems were encountered because the default .gitignore file skips files
in these directories. To work around this, separate .gitignore files were
added to subdirectories in tests/data/devuan_support/

It was also realised that some of the files in /var/lib/lists/ were very
large, and also not required by the testing: these files were removed.

This commit also merges a number of bugfixes, workarounds and skips for
pylint complaints, and changes to .gitignore files.

Also included is a change to version.py - using version number 2.2.53-1.1
(for version with devuan compatibility) broke several tests. So reverted
to 2.2.53.